Error message when invoked outside a pack directory
Some gift commands require that the user's present directory be a
pack's root directory.  The old code silently failed back to the help
command in this circumstance.  Now we generate an informative message
instructing the user how to resolve the problem.

I'm slightly hesitant about halting the program from
inside_a_pack_directory.  Throwing an exception might be a better
approach.  It's easy enough to change later if I'm wrong.