Make gift work with SWI Prolog 6.5.2
That version changes the way that command line arguments are handled.
The best practice is to use flags to specify desired behavior, rather
than the command line.  Unfortunately, there doesn't appear to be a
way to use flags to prevent the user's initialization file from
loading.  The "-f none" fix doesn't work on Linux.