One module per gift command
This makes it clearer which code belongs to which command.  It should
make gift load faster on the command line because it will only load code
for the command the user requests.  It also gives users a means to
extend the gift command set by creating gift_command_foo modules.

There's slightly more code overhead per command under this approach, but
I think the costs are worth the benefits.